requirements-lambda.txt
boto3==1.34.34
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
pytz==2024.1
